requirements.txt
fastapi
uvicorn 
google-generativeai
pillow
python-multipart